msdk: enc: retire tasks in order only
The get_free_task() function could let the encoder retire tasks out of
order if it had decided to skip around in the task list. This was never
happening, but if tasks do get retired out of order then the generated
bitstream would be invalid. Instead, remove this function so that no
code exists that could retire tasks out of order.