msdk: enc: defer encoder initialization until a frame comes
Currently the plugin decides whether a frame copy to fixup strides
is required when the input caps are negotiated, however this is
too early to know if a copy is required. Split encoder
initialization into two steps.  The mfxVideoParam gets initialized
when we get sink caps in set_format. The actual encoder
initialization is then deferred until the arrival of the first
frame. With this sequencing of initialization it would be possible
in a later patch to make the decision about the necessity of the
aligning copy using the actual input buffer.